chore: rename modifier, extend abstract class
