+++ model/Form/ServiceProvider/FormServiceProvider.php
@@ -1,13 +1,35 @@
 <?php
 
+/**
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; under version 2
+ * of the License (non-upgradable).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ * Copyright (c) 2024 (original work) Open Assessment Technologies SA.
+ */
+
+declare(strict_types=1);
+
 namespace oat\taoQtiItem\model\Form\ServiceProvider;
 
 use oat\generis\model\data\Ontology;
 use oat\generis\model\DependencyInjection\ContainerServiceProviderInterface;
+use oat\tao\model\featureFlag\FeatureFlagChecker;
 use oat\tao\model\form\Modifier\FormModifierManager;
-use oat\taoQtiItem\model\Form\Modifier\EditTranslationInstanceFormModifier;
+use oat\taoQtiItem\model\Form\Modifier\TranslationInstanceFormModifier;
 use oat\taoQtiItem\model\qti\Service;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 
 class FormServiceProvider implements ContainerServiceProviderInterface
@@ -17,10 +39,11 @@ class FormServiceProvider implements ContainerServiceProviderInterface
         $services = $configurator->services();
 
         $services
-            ->set(EditTranslationInstanceFormModifier::class, EditTranslationInstanceFormModifier::class)
+            ->set(TranslationInstanceFormModifier::class, TranslationInstanceFormModifier::class)
             ->args([
                 service(Ontology::SERVICE_ID),
                 service(Service::class),
+                service(FeatureFlagChecker::class),
             ]);
 
         $formModifierManager = $services->get(FormModifierManager::class);
@@ -28,8 +51,8 @@ class FormServiceProvider implements ContainerServiceProviderInterface
             ->call(
                 'add',
                 [
-                    service(EditTranslationInstanceFormModifier::class),
-                    EditTranslationInstanceFormModifier::ID,
+                    service(TranslationInstanceFormModifier::class),
+                    TranslationInstanceFormModifier::ID,
                 ]
             );
     }
